requirements.txt
python-json-logger==2.0.4
paho-mqtt==2.1.0